+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Flowchart subgraph support.
+
+### Fixed
+
+- Flowchart nodes with shape but without text use key as default.
+
 ## [0.4.0] - 2023-09-10
 
 ### Added